chore: clean up PR #73 churn from scheduling round-trip
After scheduling moved back into core, several artifacts from the brief
"scheduling as a module" iteration remained as pure noise. Cleanup:

- Drop migrations 0004/0005/0006 (FK drop + schedule_name add/drop +
  package_schedules recreate). Net effect on a fresh install was zero;
  the branch was never released so no install ever applied them.
- Revert services/scheduling.ts back to scheduler.ts (cosmetic rename
  with no functional value).
- Re-merge packages/db/src/schema/schedules.ts back into runs.ts where
  packageSchedules originally lived.
- Restore the Schedule component schema to its original position in
  openapi/schemas.ts (it had been moved ~100 lines up).
- Restore the leading blank line in openapi/paths/schedules.ts.
- Drop the enrichRun merge hook infrastructure (YAGNI). Since scheduling
  is back in core, no module implements it. Removes EnrichRunInput,
  enrichRuns, callMergeHook, the dedicated test, and the call sites in
  state/runs.ts. The hook system keeps callHook (first-match) and
  emitEvent (broadcast) — adding merge semantics back is a small change
  if a future module ever needs it.

Verified: bun run check passes; full test suite (1923 tests) passes.
