+++ CLAUDE.md
@@ -162,10 +162,10 @@ Appstrate uses a formalized module system for optional features. The contract is
 **Key files:**
 
 - `packages/core/src/module.ts` — `AppstrateModule` interface, `ModuleInitContext`, hook & event type maps (framework-agnostic, published on npm)
-- `apps/api/src/lib/modules/module-loader.ts` — Loader with built-in auto-discovery, dynamic import, topological sort, hook/event dispatch (including merge hooks), AppConfig extension, shutdown
+- `apps/api/src/lib/modules/module-loader.ts` — Loader with built-in auto-discovery, dynamic import, topological sort, hook/event dispatch, AppConfig extension, shutdown
 - `apps/api/src/lib/modules/migrate.ts` — `applyModuleMigrations()` helper for module-owned Drizzle migrations (PostgreSQL + PGlite, per-module tracking tables, advisory-lock serialization)
 - `apps/api/src/lib/modules/registry.ts` — `getModuleRegistry()` reads `APPSTRATE_MODULES` env var (comma-separated specifiers), `buildModuleInitContext()` provides platform services
-- `apps/api/src/lib/modules/hooks.ts` — Lifecycle helpers (`beforeRun`, `onRunStatusChange`, `onOrgCreate`, `onOrgDelete`, `enrichRuns`) — call hooks by name, never by module ID
+- `apps/api/src/lib/modules/hooks.ts` — Lifecycle helpers (`beforeRun`, `onRunStatusChange`, `onOrgCreate`, `onOrgDelete`) — call hooks by name, never by module ID
 - `apps/api/src/modules/README.md` — Authoring guide for built-in modules, plus per-module READMEs next to each `index.ts`
 
 **Built-in module auto-discovery:** The loader scans `apps/api/src/modules/*/index.ts` at boot via `discoverBuiltinModules()`. Each subdirectory with an `index.ts` is registered as a built-in module with id equal to the directory name. The list is never hardcoded — adding a new module is as simple as dropping a new directory under `apps/api/src/modules/`. The same discovery is reused by `scripts/verify-openapi.ts`, so OpenAPI validation picks up new modules with zero ceremony.
@@ -179,21 +179,20 @@ Appstrate uses a formalized module system for optional features. The contract is
 | webhooks            | `webhooks`, `webhook_deliveries`  |
 | provider-management | `org_provider_keys`, `org_models` |
 
-Scheduling lives in core (`apps/api/src/services/scheduling.ts`, `apps/api/src/routes/schedules.ts`, `package_schedules` table in `packages/db/src/schema/schedules.ts`). It was briefly extracted as a module during the `feat/platform-modules` iteration, then moved back once it became clear that the coupling with `runs` (FK, filtering, enrichment, realtime) made module isolation cost more than it delivered. Webhooks and provider-management remain modules because they keep a clean boundary with core.
+Scheduling lives in core (`apps/api/src/services/scheduler.ts`, `apps/api/src/routes/schedules.ts`, `package_schedules` table in `packages/db/src/schema/runs.ts`). It was briefly extracted as a module during the `feat/platform-modules` iteration, then moved back once it became clear that the coupling with `runs` (FK, filtering, enrichment, realtime) made module isolation cost more than it delivered. Webhooks and provider-management remain modules because they keep a clean boundary with core.
 
 **FK direction rule:** Backward references (module → core) use Drizzle `.references()` inline in the module schema — safe because core tables always exist before any module migration runs. Forward references (core → module) are impossible to express via Drizzle without leaking the module schema into core, so if a future module ever needs one it must add it via raw SQL inside its own migration. Core schemas never reference module-owned tables.
 
-**Hooks vs Events vs Merge hooks:**
+**Hooks vs Events:**
 
 - **Hooks** (`callHook`): First-match-wins. Naming: `beforeX` (gates), `resolveX` (data resolution). Example: `beforeRun` blocks a run with a structured error; `resolveModel` returns the LLM model for a run.
-- **Merge hooks** (`callMergeHook`): All-match, shallow merge of partial maps keyed by row id. Used for read-time enrichment where multiple modules may contribute different fields to the same response. Example: `enrichRun` — cloud could return `{ [runId]: { billedAmount } }`, a future analytics module could return `{ [runId]: { pageViews } }`, and core merges them into the run row before serialization. Batched — core passes all rows at once so modules can do a single `IN (...)` lookup.
 - **Events** (`emitEvent`): Broadcast-to-all, side effects only. Naming: `onX` (something happened, modules react). Example: `onRunStatusChange`, `onOrgCreate`, `onOrgDelete`. Errors in individual handlers are isolated.
 
 The platform calls hooks/events by name, never by module ID. This ensures zero knowledge of module internals.
 
 **Permissions & API key scopes:** RBAC is a platform capability, not a module concern. Core's `apps/api/src/lib/permissions.ts` is the single typed source of truth — it ships the full `Permission` union (including `webhooks:*`, `models:*`, `provider-keys:*`), the role-to-permission matrix, and the API key allowlist. Module manifests do not declare `permissions` or `apiKeyScopes`. Module routes protect handlers with the same typed `requirePermission("webhooks", "write")` helper core uses, so adding a new resource requires editing `permissions.ts` and the module in the same PR. Disabling a module leaves inert permission strings in the role sets (the routes that read them are not loaded), which is a deliberate trade-off for keeping the type system honest.
 
-**Creating a new built-in module:** Drop a directory under `apps/api/src/modules/<name>/` with an `index.ts` exporting a default `AppstrateModule`. Discovery picks it up automatically. The module contributes feature flags via `features`, routes via `createRouter()`, auth-bypass paths via `publicPaths`, email template overrides via `emailOverrides`, request/response logic via `hooks` (`beforeRun`, `beforeSignup`, `resolveModel`, `enrichRun`), notifications via `events` (`onRunStatusChange`, `onOrgCreate`, `onOrgDelete`), and database tables via module-owned Drizzle migrations. If the module introduces a new RBAC resource, extend `apps/api/src/lib/permissions.ts` in the same PR. To ship as an external npm package instead, export the same default and set `APPSTRATE_MODULES=@scope/name`.
+**Creating a new built-in module:** Drop a directory under `apps/api/src/modules/<name>/` with an `index.ts` exporting a default `AppstrateModule`. Discovery picks it up automatically. The module contributes feature flags via `features`, routes via `createRouter()`, auth-bypass paths via `publicPaths`, email template overrides via `emailOverrides`, request/response logic via `hooks` (`beforeRun`, `beforeSignup`, `resolveModel`), notifications via `events` (`onRunStatusChange`, `onOrgCreate`, `onOrgDelete`), and database tables via module-owned Drizzle migrations. If the module introduces a new RBAC resource, extend `apps/api/src/lib/permissions.ts` in the same PR. To ship as an external npm package instead, export the same default and set `APPSTRATE_MODULES=@scope/name`.
 
 **Disabling a module = zero footprint:** remove it from `APPSTRATE_MODULES` and it is neither imported nor initialized. No tables, no routes, no middleware, no hook handlers, no feature flags. Core knows nothing about the module's existence. Default is empty (OSS mode).
