+++ apps/api/src/openapi/paths/schedules.ts
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
+
 export const schedulesPaths = {
   "/api/schedules": {
     get: {
