--- packages/db/drizzle/0005_drop_run_schedule_name.sql
@@ -1,3 +1,0 @@
--- Drop denormalized runs.schedule_name — modules now enrich runs via the
--- `enrichRun` hook instead. Keeps core agnostic of module-owned tables.
-ALTER TABLE "runs" DROP COLUMN IF EXISTS "schedule_name";
