+++ packages/db/drizzle/meta/_journal.json
@@ -29,27 +29,6 @@
       "when": 1775656830201,
       "tag": "0003_overrated_ozymandias",
       "breakpoints": true
-    },
-    {
-      "idx": 4,
-      "version": "7",
-      "when": 1775700000000,
-      "tag": "0004_denormalize_schedule_name",
-      "breakpoints": true
-    },
-    {
-      "idx": 5,
-      "version": "7",
-      "when": 1775800000000,
-      "tag": "0005_drop_run_schedule_name",
-      "breakpoints": true
-    },
-    {
-      "idx": 6,
-      "version": "7",
-      "when": 1775900000000,
-      "tag": "0006_schedules_into_core",
-      "breakpoints": true
     }
   ]
 }
