+++ packages/db/src/schema/index.ts
@@ -6,7 +6,6 @@ export * from "./organizations.ts";
 export * from "./applications.ts";
 export * from "./profiles.ts";
 export * from "./packages.ts";
-export * from "./connections.ts";
-export * from "./schedules.ts";
 export * from "./runs.ts";
+export * from "./connections.ts";
 export * from "./types.ts";
