+++ packages/db/src/schema/runs.ts
@@ -20,7 +20,6 @@ import { applications, endUsers } from "./applications.ts";
 import { apiKeys, organizations } from "./organizations.ts";
 import { packages } from "./packages.ts";
 import { connectionProfiles } from "./connections.ts";
-import { packageSchedules } from "./schedules.ts";
 
 export const runs = pgTable(
   "runs",
@@ -133,5 +132,39 @@ export const packageMemories = pgTable(
     index("idx_package_memories_package_app").on(table.packageId, table.applicationId),
     index("idx_package_memories_org_id").on(table.orgId),
     index("idx_package_memories_app_id").on(table.applicationId),
+  ],
+);
+
+export const packageSchedules = pgTable(
+  "package_schedules",
+  {
+    id: text("id").primaryKey(),
+    packageId: text("package_id")
+      .notNull()
+      .references(() => packages.id, { onDelete: "cascade" }),
+    connectionProfileId: uuid("connection_profile_id")
+      .notNull()
+      .references(() => connectionProfiles.id, { onDelete: "cascade" }),
+    orgId: uuid("org_id")
+      .notNull()
+      .references(() => organizations.id, { onDelete: "cascade" }),
+    applicationId: text("application_id")
+      .notNull()
+      .references(() => applications.id, { onDelete: "cascade" }),
+    name: text("name"),
+    enabled: boolean("enabled").default(true).notNull(),
+    cronExpression: text("cron_expression").notNull(),
+    timezone: text("timezone").default("UTC"),
+    input: jsonb("input"),
+    lastRunAt: timestamp("last_run_at"),
+    nextRunAt: timestamp("next_run_at"),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().notNull(),
+  },
+  (table) => [
+    index("idx_schedules_package_id").on(table.packageId),
+    index("idx_schedules_connection_profile_id").on(table.connectionProfileId),
+    index("idx_package_schedules_org_id").on(table.orgId),
+    index("idx_package_schedules_app_id").on(table.applicationId),
   ],
 );
