--- packages/db/src/schema/schedules.ts
@@ -1,41 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-import { pgTable, text, timestamp, boolean, uuid, index, jsonb } from "drizzle-orm/pg-core";
-import { organizations } from "./organizations.ts";
-import { applications } from "./applications.ts";
-import { packages } from "./packages.ts";
-import { connectionProfiles } from "./connections.ts";
-
-export const packageSchedules = pgTable(
-  "package_schedules",
-  {
-    id: text("id").primaryKey(),
-    packageId: text("package_id")
-      .notNull()
-      .references(() => packages.id, { onDelete: "cascade" }),
-    connectionProfileId: uuid("connection_profile_id")
-      .notNull()
-      .references(() => connectionProfiles.id, { onDelete: "cascade" }),
-    orgId: uuid("org_id")
-      .notNull()
-      .references(() => organizations.id, { onDelete: "cascade" }),
-    applicationId: text("application_id")
-      .notNull()
-      .references(() => applications.id, { onDelete: "cascade" }),
-    name: text("name"),
-    enabled: boolean("enabled").default(true).notNull(),
-    cronExpression: text("cron_expression").notNull(),
-    timezone: text("timezone").default("UTC"),
-    input: jsonb("input"),
-    lastRunAt: timestamp("last_run_at"),
-    nextRunAt: timestamp("next_run_at"),
-    createdAt: timestamp("created_at").defaultNow().notNull(),
-    updatedAt: timestamp("updated_at").defaultNow().notNull(),
-  },
-  (table) => [
-    index("idx_schedules_package_id").on(table.packageId),
-    index("idx_schedules_connection_profile_id").on(table.connectionProfileId),
-    index("idx_package_schedules_org_id").on(table.orgId),
-    index("idx_package_schedules_app_id").on(table.applicationId),
-  ],
-);
